Gerenciamento de Departamento 1.0

+++ trunk/projetoaula/departamento/exclusao_departamento.py
@@ -27,7 +27,13 @@ print "------ EXCLUSÃO DE DEPARTAMENTO ------"
 saida = None
 while saida <> "s":
     deletado = raw_input('Digite o id do departamento que deseja excluir: ')
-    excluirDepartamento(cursor, deletado)
-    saida = raw_input('Digite s para sair ou enter para continuar excluindo: ')
+    verificar_dep = "select id_departamento from departamentos where id_departamento='%s'" %(deletado)
+    existe = cursor.execute(verificar_dep)
+    if existe < 1:
+        print "Departamento não existente."
+        continue
+    else:
+        excluirDepartamento(cursor, deletado)
+        saida = raw_input('Digite s para sair ou enter para continuar excluindo: ')
 print "------ FINISH ------"
 db.close()
